Refactoring Summary: I created a new package ...simulator.ui.components.list.messaging and extracted the following parts: 1.  MessagingStep.kt: Now serves as a clean main orchestrator for the simulator's messaging step. 2.  UserPortraitSlot.kt: Extracted the premium identity card. It handles the user portrait display, camera/gallery actions, and the collapsible seasonal plot toggle. 3.  FindingsDialog.kt: Houses the complex "ML Face Detection Findings" dashboard, including the technical telemetry and raw output analysis. 4.  FaceTelemetryVisualizer.kt: Modularized the HUD-style canvas overlay that maps biometric landmarks onto the portrait. 5.  SeasonalQuadrantMap.kt: Isolated the 2D geometric plot for seasonal classification. 6.  ProfileTranslation.kt: Moved all color theory translation logic (Contrast, Temperature, and Depth profiles) into a single utility file. Result: The code is now significantly cleaner, more organized, and follows best practices for Compose architecture. I have also verified that the project builds successfully and the legacy monolith file has been removed.
