+++ applications/kocolor/features/analyzer/src/main/java/com/zoewave/probase/kocolor/features/analyzer/simulator/ui/StyleSimulatorScreen.kt
@@ -37,7 +37,7 @@ import com.zoewave.probase.kocolor.features.analyzer.R
 import com.zoewave.probase.kocolor.features.analyzer.simulator.ui.components.ResultStep
 import com.zoewave.probase.kocolor.features.analyzer.simulator.ui.components.graphics.AnalysisStep
 import com.zoewave.probase.kocolor.features.analyzer.simulator.ui.components.graphics.MagicBackground
-import com.zoewave.probase.kocolor.features.analyzer.simulator.ui.components.list.MessagingStep
+import com.zoewave.probase.kocolor.features.analyzer.simulator.ui.components.list.messaging.MessagingStep
 import com.zoewave.probase.kocolor.model.KoColorRoute
 
 @OptIn(ExperimentalMaterial3Api::class)
